Refactor logs implementation into EventLogsPipeline.
