+++ src/Microsoft.Diagnostics.Monitoring.EventPipe/DiagnosticsEventPipeProcessor.cs
@@ -6,11 +6,8 @@ using Graphs;
 using Microsoft.Diagnostics.NETCore.Client;
 using Microsoft.Diagnostics.Tracing;
 using Microsoft.Diagnostics.Tracing.Parsers.Clr;
-using Microsoft.Extensions.Logging;
 using System;
-using System.Collections.Generic;
 using System.IO;
-using System.Text.Json;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -19,9 +16,7 @@ namespace Microsoft.Diagnostics.Monitoring.EventPipe
     internal partial class DiagnosticsEventPipeProcessor : IAsyncDisposable
     {
         private readonly MemoryGraph _gcGraph;
-        private readonly ILoggerFactory _loggerFactory;
         private readonly PipeMode _mode;
-        private readonly LogLevel _logsLevel;
         private readonly Func<string, CancellationToken, Task> _processInfoCallback;
         private readonly MonitoringSourceConfiguration _userConfig;
         private readonly Func<Stream, CancellationToken, Task> _onStreamAvailable;
@@ -38,8 +33,6 @@ namespace Microsoft.Diagnostics.Monitoring.EventPipe
 
         public DiagnosticsEventPipeProcessor(
             PipeMode mode,
-            ILoggerFactory loggerFactory = null,              // PipeMode = Logs
-            LogLevel logsLevel = LogLevel.Debug,              // PipeMode = Logs
             MemoryGraph gcGraph = null,                       // PipeMode = GCDump
             MonitoringSourceConfiguration configuration = null, // PipeMode = Nettrace, EventSource
             Func<Stream, CancellationToken, Task> onStreamAvailable = null, // PipeMode = Nettrace
@@ -50,9 +43,7 @@ namespace Microsoft.Diagnostics.Monitoring.EventPipe
             )
         {
             _mode = mode;
-            _loggerFactory = loggerFactory;
             _gcGraph = gcGraph;
-            _logsLevel = logsLevel;
             _processInfoCallback = processInfoCallback;
             _userConfig = configuration;
             _onStreamAvailable = onStreamAvailable;
@@ -77,11 +68,7 @@ namespace Microsoft.Diagnostics.Monitoring.EventPipe
                 try
                 {
                     MonitoringSourceConfiguration config = null;
-                    if (_mode == PipeMode.Logs)
-                    {
-                        config = new LoggingSourceConfiguration(_logsLevel);
-                    }
-                    else if (_mode == PipeMode.GCDump)
+                    if (_mode == PipeMode.GCDump)
                     {
                         config = new GCDumpSourceConfiguration();
                     }
@@ -122,11 +109,6 @@ namespace Microsoft.Diagnostics.Monitoring.EventPipe
                     if (_mode == PipeMode.EventSource)
                     {
                         _onEventSourceAvailable(source);
-                    }
-                    else if (_mode == PipeMode.Logs)
-                    {
-                        // Logging
-                        HandleLoggingEvents(source);
                     }
                     else if (_mode == PipeMode.GCDump)
                     {
@@ -214,144 +196,6 @@ namespace Microsoft.Diagnostics.Monitoring.EventPipe
             {
                 await stopFunc();
             }
-        }
-
-        private void HandleLoggingEvents(EventPipeEventSource source)
-        {
-            string lastFormattedMessage = string.Empty;
-
-            var logActivities = new Dictionary<Guid, LogActivityItem>();
-            var stack = new Stack<Guid>();
-
-            source.Dynamic.AddCallbackForProviderEvent(LoggingSourceConfiguration.MicrosoftExtensionsLoggingProviderName, "ActivityJson/Start", (traceEvent) =>
-            {
-                var factoryId = (int)traceEvent.PayloadByName("FactoryID");
-                var categoryName = (string)traceEvent.PayloadByName("LoggerName");
-                var argsJson = (string)traceEvent.PayloadByName("ArgumentsJson");
-
-                // TODO: Store this information by logger factory id
-                var item = new LogActivityItem
-                {
-                    ActivityID = traceEvent.ActivityID,
-                    ScopedObject = new LogObject(JsonDocument.Parse(argsJson).RootElement),
-                };
-
-                if (stack.Count > 0)
-                {
-                    Guid parentId = stack.Peek();
-                    if (logActivities.TryGetValue(parentId, out var parentItem))
-                    {
-                        item.Parent = parentItem;
-                    }
-                }
-
-                stack.Push(traceEvent.ActivityID);
-                logActivities[traceEvent.ActivityID] = item;
-            });
-
-            source.Dynamic.AddCallbackForProviderEvent(LoggingSourceConfiguration.MicrosoftExtensionsLoggingProviderName, "ActivityJson/Stop", (traceEvent) =>
-            {
-                var factoryId = (int)traceEvent.PayloadByName("FactoryID");
-                var categoryName = (string)traceEvent.PayloadByName("LoggerName");
-
-                stack.Pop();
-                logActivities.Remove(traceEvent.ActivityID);
-            });
-
-            source.Dynamic.AddCallbackForProviderEvent(LoggingSourceConfiguration.MicrosoftExtensionsLoggingProviderName, "MessageJson", (traceEvent) =>
-            {
-                // Level, FactoryID, LoggerName, EventID, EventName, ExceptionJson, ArgumentsJson
-                var logLevel = (LogLevel)traceEvent.PayloadByName("Level");
-                var factoryId = (int)traceEvent.PayloadByName("FactoryID");
-                var categoryName = (string)traceEvent.PayloadByName("LoggerName");
-                var eventId = (int)traceEvent.PayloadByName("EventId");
-                var eventName = (string)traceEvent.PayloadByName("EventName");
-                var exceptionJson = (string)traceEvent.PayloadByName("ExceptionJson");
-                var argsJson = (string)traceEvent.PayloadByName("ArgumentsJson");
-
-                // There's a bug that causes some of the columns to get mixed up
-                if (eventName.StartsWith("{"))
-                {
-                    argsJson = exceptionJson;
-                    exceptionJson = eventName;
-                    eventName = null;
-                }
-
-                if (string.IsNullOrEmpty(argsJson))
-                {
-                    return;
-                }
-
-                Exception exception = null;
-
-                ILogger logger = _loggerFactory.CreateLogger(categoryName);
-                var scopes = new List<IDisposable>();
-
-                if (logActivities.TryGetValue(traceEvent.ActivityID, out var logActivityItem))
-                {
-                    // REVIEW: Does order matter here? We're combining everything anyways.
-                    while (logActivityItem != null)
-                    {
-                        scopes.Add(logger.BeginScope(logActivityItem.ScopedObject));
-
-                        logActivityItem = logActivityItem.Parent;
-                    }
-                }
-
-                try
-                {
-                    if (exceptionJson != "{}")
-                    {
-                        var exceptionMessage = JsonSerializer.Deserialize<JsonElement>(exceptionJson);
-                        exception = new LoggerException(exceptionMessage);
-                    }
-
-                    var message = JsonSerializer.Deserialize<JsonElement>(argsJson);
-                    if (message.TryGetProperty("{OriginalFormat}", out var formatElement))
-                    {
-                        var formatString = formatElement.GetString();
-                        var formatter = new LogValuesFormatter(formatString);
-                        object[] args = new object[formatter.ValueNames.Count];
-                        for (int i = 0; i < args.Length; i++)
-                        {
-                            args[i] = message.GetProperty(formatter.ValueNames[i]).GetString();
-                        }
-
-                        logger.Log(logLevel, new EventId(eventId, eventName), exception, formatString, args);
-                    }
-                    else
-                    {
-                        var obj = new LogObject(message, lastFormattedMessage);
-                        logger.Log(logLevel, new EventId(eventId, eventName), obj, exception, LogObject.Callback);
-                    }
-                }
-                catch (Exception)
-                {
-                }
-                finally
-                {
-                    scopes.ForEach(d => d.Dispose());
-                }
-            });
-
-            source.Dynamic.AddCallbackForProviderEvent(LoggingSourceConfiguration.MicrosoftExtensionsLoggingProviderName, "FormattedMessage", (traceEvent) =>
-            {
-                // Level, FactoryID, LoggerName, EventID, EventName, FormattedMessage
-                var logLevel = (LogLevel)traceEvent.PayloadByName("Level");
-                var factoryId = (int)traceEvent.PayloadByName("FactoryID");
-                var categoryName = (string)traceEvent.PayloadByName("LoggerName");
-                var eventId = (int)traceEvent.PayloadByName("EventId");
-                var eventName = (string)traceEvent.PayloadByName("EventName");
-                var formattedMessage = (string)traceEvent.PayloadByName("FormattedMessage");
-
-                if (string.IsNullOrEmpty(formattedMessage))
-                {
-                    formattedMessage = eventName;
-                    eventName = string.Empty;
-                }
-
-                lastFormattedMessage = formattedMessage;
-            });
         }
 
         private async Task HandleGCEvents(EventPipeEventSource source, Func<Task> stopFunc, CancellationToken token)
@@ -505,15 +349,6 @@ namespace Microsoft.Diagnostics.Monitoring.EventPipe
             }
 
             _eventPipeSession?.Dispose();
-        }
-
-        private class LogActivityItem
-        {
-            public Guid ActivityID { get; set; }
-
-            public LogObject ScopedObject { get; set; }
-
-            public LogActivityItem Parent { get; set; }
         }
     }
 }
