+++ src/Microsoft.Diagnostics.Monitoring.EventPipe/Logs/EventLogsPipeline.cs
@@ -3,12 +3,11 @@
 // See the LICENSE file in the project root for more information.
 
 using Microsoft.Diagnostics.NETCore.Client;
+using Microsoft.Diagnostics.Tracing;
 using Microsoft.Extensions.Logging;
 using System;
 using System.Collections.Generic;
-using System.Text;
-using System.Threading;
-using System.Threading.Tasks;
+using System.Text.Json;
 
 namespace Microsoft.Diagnostics.Monitoring.EventPipe
 {
@@ -21,9 +20,156 @@ namespace Microsoft.Diagnostics.Monitoring.EventPipe
             _factory = factory ?? throw new ArgumentNullException(nameof(factory));
         }
 
-        internal override DiagnosticsEventPipeProcessor CreateProcessor()
+        internal override MonitoringSourceConfiguration CreateConfiguration()
         {
-            return new DiagnosticsEventPipeProcessor(PipeMode.Logs, loggerFactory: _factory, logsLevel: Settings.LogLevel);
+            return new LoggingSourceConfiguration(Settings.LogLevel);
+        }
+
+        protected override void OnEventSourceAvailable(EventPipeEventSource eventSource)
+        {
+            string lastFormattedMessage = string.Empty;
+
+            var logActivities = new Dictionary<Guid, LogActivityItem>();
+            var stack = new Stack<Guid>();
+
+            eventSource.Dynamic.AddCallbackForProviderEvent(LoggingSourceConfiguration.MicrosoftExtensionsLoggingProviderName, "ActivityJson/Start", (traceEvent) =>
+            {
+                var factoryId = (int)traceEvent.PayloadByName("FactoryID");
+                var categoryName = (string)traceEvent.PayloadByName("LoggerName");
+                var argsJson = (string)traceEvent.PayloadByName("ArgumentsJson");
+
+                // TODO: Store this information by logger factory id
+                var item = new LogActivityItem
+                {
+                    ActivityID = traceEvent.ActivityID,
+                    ScopedObject = new LogObject(JsonDocument.Parse(argsJson).RootElement),
+                };
+
+                if (stack.Count > 0)
+                {
+                    Guid parentId = stack.Peek();
+                    if (logActivities.TryGetValue(parentId, out var parentItem))
+                    {
+                        item.Parent = parentItem;
+                    }
+                }
+
+                stack.Push(traceEvent.ActivityID);
+                logActivities[traceEvent.ActivityID] = item;
+            });
+
+            eventSource.Dynamic.AddCallbackForProviderEvent(LoggingSourceConfiguration.MicrosoftExtensionsLoggingProviderName, "ActivityJson/Stop", (traceEvent) =>
+            {
+                var factoryId = (int)traceEvent.PayloadByName("FactoryID");
+                var categoryName = (string)traceEvent.PayloadByName("LoggerName");
+
+                stack.Pop();
+                logActivities.Remove(traceEvent.ActivityID);
+            });
+
+            eventSource.Dynamic.AddCallbackForProviderEvent(LoggingSourceConfiguration.MicrosoftExtensionsLoggingProviderName, "MessageJson", (traceEvent) =>
+            {
+                // Level, FactoryID, LoggerName, EventID, EventName, ExceptionJson, ArgumentsJson
+                var logLevel = (LogLevel)traceEvent.PayloadByName("Level");
+                var factoryId = (int)traceEvent.PayloadByName("FactoryID");
+                var categoryName = (string)traceEvent.PayloadByName("LoggerName");
+                var eventId = (int)traceEvent.PayloadByName("EventId");
+                var eventName = (string)traceEvent.PayloadByName("EventName");
+                var exceptionJson = (string)traceEvent.PayloadByName("ExceptionJson");
+                var argsJson = (string)traceEvent.PayloadByName("ArgumentsJson");
+
+                // There's a bug that causes some of the columns to get mixed up
+                if (eventName.StartsWith("{"))
+                {
+                    argsJson = exceptionJson;
+                    exceptionJson = eventName;
+                    eventName = null;
+                }
+
+                if (string.IsNullOrEmpty(argsJson))
+                {
+                    return;
+                }
+
+                Exception exception = null;
+
+                ILogger logger = _factory.CreateLogger(categoryName);
+                var scopes = new List<IDisposable>();
+
+                if (logActivities.TryGetValue(traceEvent.ActivityID, out var logActivityItem))
+                {
+                    // REVIEW: Does order matter here? We're combining everything anyways.
+                    while (logActivityItem != null)
+                    {
+                        scopes.Add(logger.BeginScope(logActivityItem.ScopedObject));
+
+                        logActivityItem = logActivityItem.Parent;
+                    }
+                }
+
+                try
+                {
+                    if (exceptionJson != "{}")
+                    {
+                        var exceptionMessage = JsonSerializer.Deserialize<JsonElement>(exceptionJson);
+                        exception = new LoggerException(exceptionMessage);
+                    }
+
+                    var message = JsonSerializer.Deserialize<JsonElement>(argsJson);
+                    if (message.TryGetProperty("{OriginalFormat}", out var formatElement))
+                    {
+                        var formatString = formatElement.GetString();
+                        var formatter = new LogValuesFormatter(formatString);
+                        object[] args = new object[formatter.ValueNames.Count];
+                        for (int i = 0; i < args.Length; i++)
+                        {
+                            args[i] = message.GetProperty(formatter.ValueNames[i]).GetString();
+                        }
+
+                        logger.Log(logLevel, new EventId(eventId, eventName), exception, formatString, args);
+                    }
+                    else
+                    {
+                        var obj = new LogObject(message, lastFormattedMessage);
+                        logger.Log(logLevel, new EventId(eventId, eventName), obj, exception, LogObject.Callback);
+                    }
+                }
+                catch (Exception)
+                {
+                }
+                finally
+                {
+                    scopes.ForEach(d => d.Dispose());
+                }
+            });
+
+            eventSource.Dynamic.AddCallbackForProviderEvent(LoggingSourceConfiguration.MicrosoftExtensionsLoggingProviderName, "FormattedMessage", (traceEvent) =>
+            {
+                // Level, FactoryID, LoggerName, EventID, EventName, FormattedMessage
+                var logLevel = (LogLevel)traceEvent.PayloadByName("Level");
+                var factoryId = (int)traceEvent.PayloadByName("FactoryID");
+                var categoryName = (string)traceEvent.PayloadByName("LoggerName");
+                var eventId = (int)traceEvent.PayloadByName("EventId");
+                var eventName = (string)traceEvent.PayloadByName("EventName");
+                var formattedMessage = (string)traceEvent.PayloadByName("FormattedMessage");
+
+                if (string.IsNullOrEmpty(formattedMessage))
+                {
+                    formattedMessage = eventName;
+                    eventName = string.Empty;
+                }
+
+                lastFormattedMessage = formattedMessage;
+            });
+        }
+
+        private class LogActivityItem
+        {
+            public Guid ActivityID { get; set; }
+
+            public LogObject ScopedObject { get; set; }
+
+            public LogActivityItem Parent { get; set; }
         }
     }
 }
